[Snippets[CPU] Applied Ivan comments
